server/requirements.txt
Django==4.0.7
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-filter==22.1
pycryptodomex==3.15.0
djangorestframework-simplejwt==5.2.0
celery==5.2.7
django-redis==5.2.0
django-celery-results==2.4.0
coloredlogs==15.0.1
typing_extensions==4.3.0
requests==2.28.1
tqdm==4.64.1
django-celery-beat==2.3.0
mysqlclient==2.1.1
psutil==5.9.2
python-daemon==2.3.1
gunicorn==20.1.0
flower==1.2.0
django-proxy==1.2.1